Improve interpolation in calipso import
Use the new interpolation algorithm of ccext (interpolation: "smart").
Rename the old interpolation to "nearest-old". Add a new faster
nearest-neighbor interpolation which does not use cctk, but numpy array
indexing. This is what is now called "nearest".

Rewrite the code of the calipso import class for better clarity in the
process.